feat(settings): combined /settings/data (Export + Knowledge Base) (#231)
PRD #226 · Slice 5. Mirrors the slice-1 pattern (#227/#235):

- New `/settings/data` shell renders two tabs in order (Export, Knowledge Base)
  via the shared SettingsTabs (URL-synced ?tab=…).
- Tab bodies extracted into `export-tab.tsx` / `knowledge-base-tab.tsx` —
  content unchanged.
- Old URLs redirect: `/settings/export` → `/settings/data?tab=export`;
  `/settings/knowledge` → `/settings/data?tab=knowledge-base`.
- `settings-nav.ts` collapses the two old entries into a single "Data" entry
  (Download icon); the mobile dropdown is driven by the same array, so it
  picks up the new entry for free.

Smoke test pins the tab order; per-tab behavior is covered by the existing
API-route tests (`/api/settings/export`, `/api/knowledge/*`).

Per the PRD, Knowledge Base is expected to migrate out of Settings entirely
in a follow-up; for this slice it lives under Data as a tab.
