+++ src/lib/settings-nav.ts
@@ -10,7 +10,6 @@ import {
   Bell,
   FileText,
   Download,
-  BookOpen,
   Send,
   Link2,
   CreditCard,
@@ -31,20 +30,35 @@ export const settingsNavItems: SettingsNavItem[] = [
   { href: "/settings/appearance", label: "Appearance", icon: Palette },
   { href: "/settings/jobs", label: "Jobs", icon: ListChecks },
   { href: "/settings/item-library", label: "Item Library", icon: Library },
-  { href: "/settings/estimate-templates", label: "Estimate Templates", icon: LayoutTemplate },
+  {
+    href: "/settings/estimate-templates",
+    label: "Estimate Templates",
+    icon: LayoutTemplate,
+  },
   { href: "/settings/pdf-presets", label: "PDF Presets", icon: FileText },
   { href: "/settings/vendors", label: "Vendors", icon: Store },
-  { href: "/settings/expense-categories", label: "Expense Categories", icon: Receipt },
+  {
+    href: "/settings/expense-categories",
+    label: "Expense Categories",
+    icon: Receipt,
+  },
   { href: "/settings/users", label: "Users & Crew", icon: Users },
   { href: "/settings/email", label: "Email Accounts", icon: Mail },
-  { href: "/settings/signatures", label: "Email Signatures", icon: FileSignature },
-  { href: "/settings/contract-templates", label: "Contract Templates", icon: FileText },
+  {
+    href: "/settings/signatures",
+    label: "Email Signatures",
+    icon: FileSignature,
+  },
+  {
+    href: "/settings/contract-templates",
+    label: "Contract Templates",
+    icon: FileText,
+  },
   { href: "/settings/contracts", label: "Contracts", icon: Send },
   { href: "/settings/notifications", label: "Notifications", icon: Bell },
   { href: "/settings/accounting", label: "Accounting", icon: Link2 },
   { href: "/settings/stripe", label: "Stripe Payments", icon: CreditCard },
   { href: "/settings/payments", label: "Outgoing Emails", icon: Mail },
   { href: "/settings/reports", label: "Reports", icon: FileText },
-  { href: "/settings/export", label: "Data Export", icon: Download },
-  { href: "/settings/knowledge", label: "Knowledge Base", icon: BookOpen },
+  { href: "/settings/data", label: "Data", icon: Download },
 ];
